Remove redundant open() and close() invocations
Checking to see if the file is open before it is read (by the mmdb-reader package) is
redundant and introduces a race condition (the file permissions can be modified or 
the file can be deleted between the initial check and the actual read. Also, both the 
`fs.open()` and `fs.close()` calls are synchronous, which is not ideal for a server.